Remove delimiter boolean logic
The boolean logic in the delimiter was causing an error.  
In the original FuelPHP class, it's checking for a delimiter either 
passed into the method, or in the config.  Since the delimiter is 
now passed through the attributes array, the 'or' can't find 
a default '$delimiter' and throws a undefined variable error.  
The array get checks for the delimiter in both the attribute array
and the config array, so it's safe to remove.  